backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.3
asttokens==2.4.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.1
click==8.3.1
comet_ml==3.49.7
comm==0.1.4
configobj==5.0.9
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.6.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.4.0
distro==1.9.0
docker==7.1.0
docopt==0.6.2
dulwich==0.22.8
ecdsa==0.19.1
everett==3.1.0
executing==2.0.0
faiss-cpu==1.13.2
fastapi==0.128.0
fastjsonschema==2.21.2
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.32.3
idna==3.10
ipykernel==6.25.2
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
langchain==1.2.4
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib==3.10.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.2.0
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==2.15.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.5.8
networkx==3.4.2
numpy==1.26.3
openai==2.15.0
opencv-python==4.9.0.80
orjson==3.11.5
ormsgpack==1.12.1
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
propcache==0.3.1
protobuf==6.33.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.16.1
pyparsing==3.2.3
pypdf==6.6.0
python-box==6.1.0
python-dateutil==2.8.2
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.2
pyzmq==25.1.1
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.25.1
setuptools==78.1.0
simplejson==3.20.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.2
SQLAlchemy==2.0.45
stack-data==0.6.3
starlette==0.50.0
streamlit==1.53.0
sympy==1.13.1
tenacity==9.1.2
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.3.3
tqdm==4.67.1
traitlets==5.11.2
transformers==4.52.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
-e git+https://github.com/ultralytics/ultralytics.git@d642347f8efbfb51e3186c5a61bc163de699f9c1#egg=ultralytics
ultralytics-thop==2.0.14
urllib3==2.4.0
uuid_utils==0.13.0
uvicorn==0.40.0
virtualenv==20.32.0
watchdog==6.0.0
wcwidth==0.2.8
webencodings==0.5.1
wrapt==1.17.2
wurlitzer==3.1.1
xxhash==3.5.0
yarg==0.1.9
yarl==1.20.0
zstandard==0.25.0
